# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is Automatic Generate Invoice API Documentation",
        "title": "Automatic Generate Invoice System API Documentation",
        "contact": {},
        "version": "2.0"
    },
    "host": "backend-capstone-alta-golang-staging.up.railway.app",
    "paths": {
        "/admin/invoice": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Admin Get All Invoice User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoice"
                ],
                "summary": "Get All Invoice",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Invoice"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entities.Invoice"
                        }
                    }
                }
            }
        },
        "/admin/invoice-payment-status": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Admin can get all invoice payment status information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoice Payment Status"
                ],
                "summary": "Get All Invoice Payment Status Information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.InvoicePaymentStatus"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Admin can create invoice payment for table invoice",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoice Payment Status"
                ],
                "summary": "Create Invoice Payment Status",
                "parameters": [
                    {
                        "description": "required",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.InvoicePaymentStatus"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.InvoicePaymentStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entities.InvoicePaymentStatus"
                        }
                    }
                }
            }
        },
        "/admin/invoice-payment-status/{id}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Admin can get invoice payment status information by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoice Payment Status"
                ],
                "summary": "Get Invoice Payment Status Information by Id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.InvoicePaymentStatus"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/entities.InvoicePaymentStatus"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "User can update invoice payment status information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoice Payment Status"
                ],
                "summary": "Update Invoice Payment Status Information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "required",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.InvoicePaymentStatus"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.InvoicePaymentStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entities.InvoicePaymentStatus"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Admin can delete invoice payment status information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoice Payment Status"
                ],
                "summary": "Delete Invoice Payment Status Information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.InvoicePaymentStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entities.InvoicePaymentStatus"
                        }
                    }
                }
            }
        },
        "/admin/invoice/{id}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Admin can Get Invoice User By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoice"
                ],
                "summary": "Get Invoice User By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Invoice"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entities.Invoice"
                        }
                    }
                }
            }
        },
...
        "/admin/send/email": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Admin can send email to customer for their invoice",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Send Customer"
                ],
                "summary": "Send Email to Customer",
                "parameters": [
                    {
                        "description": "required",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.SendCustomer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.SendCustomer"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entities.SendCustomer"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.Invoice": {
            "type": "object",
            "properties": {
                "buyer_address": {
                    "type": "string"
                },
                "buyer_email": {
                    "type": "string"
                },
                "buyer_name": {
                    "type": "string"
                },
                "buyer_phone": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "id_csv_file": {
                    "type": "integer"
                },
                "id_payment_status": {
                    "type": "integer"
                },
                "issuer_address": {
                    "type": "string"
                },
                "issuer_email": {
                    "type": "string"
                },
                "issuer_name": {
                    "type": "string"
                },
                "issuer_phone": {
                    "type": "string"
                },
                "tax": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "entities.InvoicePaymentStatus": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "entities.SendCustomer": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}